requirements.txt
python-binance
ccxt
ta-lib
plotly